requirements.txt
asgiref==3.2.10
autopep8==1.6.0
certifi==2021.5.30
dj-database-url==0.5.0
Django @ file:///tmp/build/80754af9/django_1601563330930/work
django-cors-headers==3.10.0
django-countries==7.2.1
django-debug-toolbar==3.2.2
django-heroku==0.3.1
djangorestframework @ file:///home/conda/feedstock_root/build_artifacts/djangorestframework_1616757768152/work
djangorestframework-jwt==1.11.0
gunicorn==20.1.0
numpy==1.21.4
opencv-python==4.5.4.60
Pillow==8.4.0
psycopg2 @ file:///tmp/build/80754af9/psycopg2_1594305089395/work
psycopg2-binary==2.9.3
pycodestyle==2.8.0
PyJWT==1.7.1
python-dotenv==0.19.2
pytz==2020.1
Serializer==0.2.1
simplejson==3.17.5
sqlparse @ file:///tmp/build/80754af9/sqlparse_1602184451250/work
toml==0.10.2
whitenoise==6.0.0
